requirements.txt
absl-py==2.1.0
accelerate==1.0.1
alpaca_eval==0.6.6
anthropic==0.14.0
bitsandbytes==0.49.1
datasets==3.3.2
evaluate==0.4.3
fire==0.6.0
huggingface-hub==0.24.6
numpy==1.24.3
openai==1.65.1
omegaconf==2.3.0
pandas==2.2.2
peft==0.12.0
python-dotenv==1.0.1
regex==2024.7.24
requests==2.32.3
rouge_score==0.1.2
scikit-learn==1.6.1
scipy==1.14.1
sentencepiece==0.2.0
tokenizers==0.20.3
torch==2.10.0
# for blackwell gpu eg pro 6000
# pre cuda13.0  torch2.11dev
tqdm==4.66.5
transformers==4.46.2
wandb==0.24.0
xxhash==3.5.0
hydra-core==1.3.2
tensor-parallel==1.2.4
beautifulsoup4==4.12.2